requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.5
django-cors-headers==3.12.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.15.0
idna==3.3
kombu==5.2.4
msgpack==1.0.4
packaging==21.3
prompt-toolkit==3.0.29
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==3.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
